refactor(#25): domain -> features 리네이밍
