+++ src/services/api/book.ts
@@ -1,5 +1,5 @@
 import { get } from '@/lib/HTTPClient';
-import { BookResponse, BookTitle } from '@/types/domain/book';
+import { BookResponse, BookTitle } from '@/types/features/book';
 
 const API_URL = '/api/books/title';
 const AUTH_TOKEN = `KakaoAK ${process.env.NEXT_PUBLIC_KAKAO_REST_API_KEY}`;
